All functions at root of techman
