+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = techman
+version = 1.13.2
+author = Techman55
+author_email = techman@techmandev.com
+description = General Purpose Python Library
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://py.techmandev.com
+project_urls =
+    Bug Tracker = https://github.com/Techman55/techman.py/issues
+classifiers =
+    Programming Language :: Python :: 3
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.6
+
+[options.packages.find]
+where = src
